SettingsWidget: Replaced pykdeuic4 generated code with cleaned up original ui file. Merged with SettingsWidget class and moved everything to settinswidget.py
